move TrieNode to own file and additional cleanup/review comments
