+++ src/test/e2e/differentialSync.test.ts
@@ -11,37 +11,37 @@ import { SyncEngine } from '~/network/sync/syncEngine';
 import { Factories } from '~/test/factories';
 
 const serverDb = jestRocksDB('rpcSync.test.server');
-const serverEngine = new Engine(serverDb);
-let serverSyncEngine: SyncEngine;
+const hubAStorageEngine = new Engine(serverDb);
+let hubASyncEngine: SyncEngine;
 
 const clientDb = jestRocksDB('rpcSync.test.client');
-const clientEngine = new Engine(clientDb);
-let clientSyncEngine: SyncEngine;
+const hubBStorageEngine = new Engine(clientDb);
+let hubBSyncEngine: SyncEngine;
 
 class mockRPCHandler implements RPCHandler {
   getUsers(): Promise<Set<number>> {
-    return serverEngine.getUsers();
+    return hubAStorageEngine.getUsers();
   }
   getAllCastsByUser(fid: number): Promise<Set<Cast>> {
-    return serverEngine.getAllCastsByUser(fid);
+    return hubAStorageEngine.getAllCastsByUser(fid);
   }
   getAllSignerMessagesByUser(fid: number): Promise<Set<SignerMessage>> {
-    return serverEngine.getAllSignerMessagesByUser(fid);
+    return hubAStorageEngine.getAllSignerMessagesByUser(fid);
   }
   getAllReactionsByUser(fid: number): Promise<Set<Reaction>> {
-    return serverEngine.getAllReactionsByUser(fid);
+    return hubAStorageEngine.getAllReactionsByUser(fid);
   }
   getAllFollowsByUser(fid: number): Promise<Set<Follow>> {
-    return serverEngine.getAllFollowsByUser(fid);
+    return hubAStorageEngine.getAllFollowsByUser(fid);
   }
   getAllVerificationsByUser(fid: number): Promise<Set<Verification>> {
-    return serverEngine.getAllVerificationsByUser(fid);
+    return hubAStorageEngine.getAllVerificationsByUser(fid);
   }
   getCustodyEventByUser(fid: number): Promise<Result<IdRegistryEvent, FarcasterError>> {
-    return serverEngine.getCustodyEventByUser(fid);
+    return hubAStorageEngine.getCustodyEventByUser(fid);
   }
   getSyncMetadataByPrefix(prefix: string): Promise<Result<NodeMetadata, FarcasterError>> {
-    const nodeMetadata = serverSyncEngine.getNodeMetadata(prefix);
+    const nodeMetadata = hubASyncEngine.getNodeMetadata(prefix);
     if (nodeMetadata) {
       return Promise.resolve(ok(nodeMetadata));
     } else {
@@ -49,18 +49,18 @@ class mockRPCHandler implements RPCHandler {
     }
   }
   getSyncIdsByPrefix(prefix: string): Promise<Result<string[], FarcasterError>> {
-    return Promise.resolve(ok(serverSyncEngine.getIdsByPrefix(prefix)));
+    return Promise.resolve(ok(hubASyncEngine.getIdsByPrefix(prefix)));
   }
   getMessagesByHashes(hashes: string[]): Promise<Message[]> {
-    return serverEngine.getMessagesByHashes(hashes);
+    return hubAStorageEngine.getMessagesByHashes(hashes);
   }
   async submitMessage(message: Message): Promise<Result<void, FarcasterError>> {
-    return await serverEngine.mergeMessage(message);
+    return await hubAStorageEngine.mergeMessage(message);
   }
 }
 
-let server: RPCServer;
-let client: RPCClient;
+let hubARPCServer: RPCServer;
+let hubARPCClient: RPCClient;
 
 const NUM_USERS = 5;
 const TEST_TIMEOUT = 2 * 60 * 1000; // 2 min timeout
@@ -69,19 +69,19 @@ describe('differentialSync', () => {
   let userInfos: UserInfo[];
 
   beforeEach(async () => {
-    await serverEngine._reset();
-    await clientEngine._reset();
-    serverSyncEngine = new SyncEngine(serverEngine);
-    clientSyncEngine = new SyncEngine(clientEngine);
+    await hubAStorageEngine._reset();
+    await hubBStorageEngine._reset();
+    hubASyncEngine = new SyncEngine(hubAStorageEngine);
+    hubBSyncEngine = new SyncEngine(hubBStorageEngine);
     // setup the rpc server and client
-    server = new RPCServer(new mockRPCHandler());
-    await server.start();
-    expect(server.address).not.toBeNull();
+    hubARPCServer = new RPCServer(new mockRPCHandler());
+    await hubARPCServer.start();
+    expect(hubARPCServer.address).not.toBeNull();
 
-    client = new RPCClient(server.address as AddressInfo);
+    hubARPCClient = new RPCClient(hubARPCServer.address as AddressInfo);
 
     // TODO: Ignore reactions until https://github.com/farcasterxyz/hub/issues/178 is fixed
-    userInfos = await populateEngine(serverEngine, NUM_USERS, {
+    userInfos = await populateEngine(hubAStorageEngine, NUM_USERS, {
       Verifications: 1,
       Casts: 10,
       Follows: 50,
@@ -90,39 +90,41 @@ describe('differentialSync', () => {
   }, TEST_TIMEOUT);
 
   afterEach(async () => {
-    await server.stop();
+    await hubARPCServer.stop();
   });
 
   const ensureEnginesEqual = async () => {
-    const userIds = await serverEngine.getUsers();
-    await expect(clientEngine.getUsers()).resolves.toEqual(userIds);
+    const userIds = await hubAStorageEngine.getUsers();
+    await expect(hubBStorageEngine.getUsers()).resolves.toEqual(userIds);
 
-    expect(serverSyncEngine.trie.items).toEqual(clientSyncEngine.trie.items);
-    expect(serverSyncEngine.trie.rootHash).toEqual(clientSyncEngine.trie.rootHash);
+    // If the root hash of the sync engines match, then all the child nodes should match
+    expect(hubASyncEngine.trie.items).toEqual(hubBSyncEngine.trie.items);
+    expect(hubASyncEngine.trie.rootHash).toEqual(hubBSyncEngine.trie.rootHash);
+
     for (const user of userIds) {
-      const casts = await serverEngine.getAllCastsByUser(user);
-      await expect(clientEngine.getAllCastsByUser(user)).resolves.toEqual(casts);
-      const follows = await serverEngine.getAllFollowsByUser(user);
-      await expect(clientEngine.getAllFollowsByUser(user)).resolves.toEqual(follows);
-      const reactions = await serverEngine.getAllReactionsByUser(user);
-      await expect(clientEngine.getAllReactionsByUser(user)).resolves.toEqual(reactions);
-      const verifications = await serverEngine.getAllVerificationsByUser(user);
-      await expect(clientEngine.getAllVerificationsByUser(user)).resolves.toEqual(verifications);
+      const casts = await hubAStorageEngine.getAllCastsByUser(user);
+      await expect(hubBStorageEngine.getAllCastsByUser(user)).resolves.toEqual(casts);
+      const follows = await hubAStorageEngine.getAllFollowsByUser(user);
+      await expect(hubBStorageEngine.getAllFollowsByUser(user)).resolves.toEqual(follows);
+      const reactions = await hubAStorageEngine.getAllReactionsByUser(user);
+      await expect(hubBStorageEngine.getAllReactionsByUser(user)).resolves.toEqual(reactions);
+      const verifications = await hubAStorageEngine.getAllVerificationsByUser(user);
+      await expect(hubBStorageEngine.getAllVerificationsByUser(user)).resolves.toEqual(verifications);
     }
   };
 
   test(
     'syncs data to a completely empty client',
     async () => {
-      // sanity test that the server has something in it and the client, nothing
-      const users = await serverEngine.getUsers();
+      const users = await hubAStorageEngine.getUsers();
       expect(users.size).toEqual(NUM_USERS);
-      expect(await clientEngine.getUsers()).toEqual(new Set());
+      expect(await hubBStorageEngine.getUsers()).toEqual(new Set());
 
-      const snapshot = serverSyncEngine.snapshot;
-      await clientSyncEngine.performSync(snapshot.excludedHashes, client);
+      const snapshot = hubASyncEngine.snapshot;
+      expect(hubBSyncEngine.shouldSync(snapshot.excludedHashes, snapshot.numMessages)).toBeTruthy();
+      await hubBSyncEngine.performSync(snapshot.excludedHashes, hubARPCClient);
       await ensureEnginesEqual();
-      expect(clientSyncEngine.shouldSync(snapshot.excludedHashes, snapshot.numMessages)).toBeFalsy();
+      expect(hubBSyncEngine.shouldSync(snapshot.excludedHashes, snapshot.numMessages)).toBeFalsy();
     },
     TEST_TIMEOUT
   );
@@ -130,9 +132,9 @@ describe('differentialSync', () => {
   test(
     'syncs only new data to a client with existing data',
     async () => {
-      await clientSyncEngine.performSync(serverSyncEngine.snapshot.excludedHashes, client);
+      await hubBSyncEngine.performSync(hubASyncEngine.snapshot.excludedHashes, hubARPCClient);
       const mergedHashes: string[] = [];
-      clientEngine.on('messageMerged', async (_fid, _type, message: Message) => {
+      hubBStorageEngine.on('messageMerged', async (_fid, _type, message: Message) => {
         mergedHashes.push(message.hash);
       });
 
@@ -150,31 +152,31 @@ describe('differentialSync', () => {
           },
         }
       );
-      const res = await Promise.all(serverEngine.mergeMessages(messages));
+      const res = await Promise.all(hubAStorageEngine.mergeMessages(messages));
       expect(res.every((r) => r.isOk())).toBeTruthy();
 
-      const serverSnapshot = serverSyncEngine.snapshot;
-      expect(serverSnapshot.numMessages).toBeGreaterThanOrEqual(clientSyncEngine.snapshot.numMessages);
-      expect(clientSyncEngine.shouldSync(serverSnapshot.excludedHashes, serverSnapshot.numMessages)).toBeTruthy();
+      const serverSnapshot = hubASyncEngine.snapshot;
+      expect(serverSnapshot.numMessages).toEqual(hubBSyncEngine.snapshot.numMessages + newMessages);
+      expect(hubBSyncEngine.shouldSync(serverSnapshot.excludedHashes, serverSnapshot.numMessages)).toBeTruthy();
 
-      const divergencePrefix = clientSyncEngine.trie.getDivergencePrefix(
-        clientSyncEngine.snapshot.prefix,
+      const divergencePrefix = hubBSyncEngine.trie.getDivergencePrefix(
+        hubBSyncEngine.snapshot.prefix,
         serverSnapshot.excludedHashes
       );
-      // The point of divergence should be some distance from the root, and must have few messages than the root
+      // The point of divergence should be some distance from the root, and must have fewer messages than the root
       expect(divergencePrefix.length).toBeGreaterThanOrEqual(3);
-      const serverDivergedNode = serverSyncEngine.trie.getNodeMetadata(divergencePrefix);
-      expect(serverDivergedNode).toBeDefined();
-      expect(serverDivergedNode?.numMessages).toBeLessThanOrEqual(serverSyncEngine.trie.root.items);
+      const serverDivergedNodeMetadata = hubASyncEngine.trie.getNodeMetadata(divergencePrefix);
+      expect(serverDivergedNodeMetadata).toBeDefined();
+      expect(serverDivergedNodeMetadata?.numMessages).toBeLessThanOrEqual(hubASyncEngine.trie.root.items);
 
-      await clientSyncEngine.performSync(serverSnapshot.excludedHashes, client);
+      await hubBSyncEngine.performSync(serverSnapshot.excludedHashes, hubARPCClient);
 
       await ensureEnginesEqual();
       for (const message of messages) {
         expect(mergedHashes).toContain(message.hash);
       }
       // In the worst case, the maximum number of messages merged should be the number of items under the divergence prefix
-      expect(mergedHashes.length).toBeLessThanOrEqual(serverDivergedNode?.numMessages || -1);
+      expect(mergedHashes.length).toBeLessThanOrEqual(serverDivergedNodeMetadata?.numMessages || -1);
       // Ideally the merged hashes length should be exactly equal to the new messages,
       // but, depending on the location of divergence, it's possible that some existing messages are syncd
       // So just ensure it's below some reasonable threshold
